+++ admin/view/chinh-sach/list.php
@@ -1,11 +1,5 @@
 <?php
-	$fields = [];
-    $sorts = ["id" => "DESC"];
-    $limits = [];
-    $condition = [];
-    $forms = [];
-    $search = [];
-    $data_list = $query->DanhSach("chinhsach", $fields, $condition, $sorts, $limits, $forms, $search);
+    $data_list = $query->DanhSach("chinhsach", [], [], ["id" => "DESC"], [],[], []);
 ?>
 <div class="blog medium">
 
@@ -36,7 +30,7 @@
                     <td><img src="../uploads/chinh-sach/<?=$value->hinh?>" height="40" /></td>
                     <td class="can-giua">
                         <a href="chinh-sach/edit?id=<?=$value->id?>"><i class="fal fa-edit"></i></a>
-                        <a style="cursor: pointer" data-id="<?=$value->id ?>" class="remove_chinhsach"><i class="fal fa-trash-alt"></i></a>
+                        <a onclick="confirm('Bạn có chắc muốn xóa?')"  href="chinh-sach/del?id=<?=$value->id?>" ><i class="fal fa-trash-alt"></i></a>
                     </td>
                 </tr>
                 <?php
@@ -46,22 +40,3 @@
         </tbody>
     </table>
 </div>
-<script>
-    $('.remove_chinhsach').click(function(){
-         const cfrm = confirm('Bạn có chắc chắn muốn xóa ?');
-        var id = $(this).data('id');
-        if(cfrm ==true){
-            $.ajax({
-                url : "chinh-sach/del",
-                method :"GET",
-                data:{
-                    id :id 
-                },
-                success:function(data){
-                    $('#remove'+id).remove();
-                }
-
-            })
-        }
-    });
-</script>
