+++ admin/view/nhom/add.php
@@ -1,9 +1,4 @@
 <?php
-	#Nhóm - Trang - Phân quyền
-    // require_once "model/PhanQuyen.php";
-    // $phanquyen = new PhanQuyen();
-    // $data_phanquyen = $phanquyen->NhomTrangQuyen($__NHOM__, 3);
-    // if( empty($data_phanquyen) || $data_phanquyen->sua == 0 ) header("location:./");
 
 	if(isset($_POST['submit']))
 	{
