+++ admin/view/trang/del.php
@@ -1,18 +1,6 @@
 <?php
-	#Nhóm - Trang - Phân quyền
-    // require_once "model/PhanQuyen.php";
-    // $phanquyen = new PhanQuyen();
-    // $data_phanquyen = $phanquyen->NhomTrangQuyen($__NHOM__, 1);
-    // if( empty($data_phanquyen) || $data_phanquyen->xoa == 0 )
-    // {
-    // 	header("location:./");
-    // }
-    // else
-    // {
-		// require_once "model/Trang.php";
-		// $trang = new Trang();
-		isset($_GET['id']) ? $id = $_GET['id'] : $id = 0;
-		$query->Xoa('trang',["id"=>"="],["id"=>$id]);
-		header("location:list");
-	// }
-?>
+
+	isset($_GET['id']) ? $id = $_GET['id'] : $id = 0;
+	$query->Xoa('trang',["id"=>"="],["id"=>$id]);
+	header("location:list");
+    ?>
