+++ backup/minh_sql.txt
@@ -1,0 +1,1 @@
+ALTER TABLE `phanquyen` ADD `them` INT NOT NULL AFTER `xem`;
